Colourise segwit adoption table
Set the background of the "yes" and "no" cells in the segwit adoption
page to make it easy to see which ones are done. Uses liquid scripting
and raw HTML since kramdown doesn't offer a way of applying an attribute
to a <td> element; this in turn means moving the data into a CSV file.
Yellow and blue chosen for good and bad for colour-blind compatability.