Update README.md
Improved some of the wording, consistency (e.g., changed the only instance of "pullreq" to match the multiple instances of "pull request"), punctuation, etc.  Removed a few words that weren't needed, and an extra blank line at the end of the document.  Also improved the point-form list, and implemented a small variety of other minor cosmetic improvements.