Simplify javascript
* Deletes unused js modules:
  * magnific-popup (tool for better image galleries) is not used at all
  * modernizr is not used at all (it creates window.Modernizr js object which is never used anywhere else)

* Deletes obsolete js modules:
  * respond and html5shiv are used only in IE9, which had last version in 2011. We can reasonably assume those will not be very frequent, and the website should work there as-is anyway.

* Puts javascripts to a single file
  * instead of using pre-built minified file, jekyll puts the files together through `{% include %}`

* Uses "local" javascript instead of hardlink to bitcoincore
  * This allows easier future experimentation

I wanted to also delete fitvids, but it is used on one page - https://bitcoincore.org/en/2016/01/13/development-visualisation-2015/ . So it is probably better to keep using it.

This PR also removes minification as a by-product of using jekyll for putting js together. I think the performance difference is negligible, since jquery is already minified.