Blog posts: set default permalink URL for new posts
Currently each post needs to have its URL manually specified, but all of
the information we use to construct that URL (date + short title) is
available in its file name.  This commit sets the default permalink URL
using that metadata; it can still be overridden by manually putting a
"permalink:" entry in the post's YAML header.

The benefit of this change is that it's one fewer date that needs to be
manually set by somebody on release day.

I tested that this change does not change the URLs of any existing
pages.