Site build: move build/test commands from travis to make
Having the commands in .travis.yml is ok when there's only a couple
simple commands, but as the number and complexity of the commands
increases, it's nice to have a way to quickly run them from a local
shell.

This commit moves the existing commands as-is to the 'build' and 'test'
make targets, which are themselves both run by the default 'all' target.
An additional target of 'preview' is added with a variant on the build
command for local previews.