RPC doc generating script: always connect to regtest
The recommended network for generating RPC docs is regtest, as stated in the
script's file header comments:
"// (2) install bitcoin core, set it up to use regtest"

To achieve this, the user previosly needed to manually put `regtest=1` into
`~/.bitcoin/bitcoin.conf` -- the script didn't specify a concrete network, so
without config file setting, mainnet was used. With the change in this commit,
we will always connect to regtest by passing the `-regtest` option to the
`bitcoin-cli` invokation.