Revert "Remove bitcoin.pdf" take 2
This reverts commit c131308cb9cf3538ba9db04a3002a76e6f334f62.

The original revert commit in
https://github.com/bitcoin-core/bitcoincore.org/pull/744
was closed without any reasoning even though it articulated well
in regards to permissive licensing and why it is ok for bitcoin.pdf
to be served from bitcoincore.org.

The "Remove bitcoin.pdf" in
https://github.com/bitcoin-core/bitcoincore.org/pull/740
did not provide any legal grounds for removing the PDF except for
"unclear status". In fact, it questions "unless anyone can point to an
explicit place where Satoshi licensed it under a free license" which
has been pointed out in
https://github.com/bitcoin-core/bitcoincore.org/pull/744.

Unless there's a lawyer present who can refute or confirm the
reasoning, I see no basis for removal of bitcoin.pdf.