Final update release key expiration date
Change the expiration to 2024. The key is no longer being used for new
releases, but letting it expire now would be inconvenient because there
may still be another 0.21.x release, and, people may still want to
verify older releases.

This should be enough for all releases signed with it to be EOL. After
that, it can be removed from the site completely.