Remove fitVids function call
This code calls a function that does not exist, resulting in a Javascript error on every page load.

fitVids is a jQuery plugin that maybe was part of the codebase at some point.

I don't think there are any videos on the site. I also think that there are modern CSS-based approaches to resizing videos that could be implemented if necessary.